trashbot/requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
attrs==20.1.0
black==20.8b1
boto3==1.15.12
botocore==1.18.12
certifi==2020.6.20
cffi==1.14.2
click==7.1.2
cryptography==3.1
decorator==4.4.2
iniconfig==1.0.1
isort==5.6.3
jmespath==0.10.0
more-itertools==8.4.0
mypy-extensions==0.4.3
packaging==20.4
passlib==1.7.2
pathspec==0.8.0
pdftools.pdfposter==0.7.post1
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
PyPDF2==1.26.0
pytest==6.0.1
python-dateutil==2.8.1
python-telegram-bot==13.0
pytz==2020.4
regex==2020.10.11
s3transfer==0.3.3
six==1.15.0
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.10
